group factories
