+++ .github/workflows/ci.yml
@@ -4,11 +4,6 @@ on:
   pull_request:
     types: [opened, synchronize, reopened]
 
-env:
-  CI: true
-  NODE_ENV: test
-  NODE_VER: 22
-
 jobs:
   lint:
     permissions:
@@ -23,8 +18,6 @@ jobs:
       - name: Start MySQL
         run: sudo /etc/init.d/mysql start
       - uses: actions/setup-node@v4
-        with:
-          node-version: ${{ env.NODE_VER }}
       - uses: actions/checkout@v6
       - run: npm install
       - name: Initialize MySQL
@@ -90,7 +83,6 @@ jobs:
           brew services start mysql@8.4
       - uses: actions/checkout@v6
       - uses: actions/setup-node@v6
-        name: Node ${{ matrix.node-version }} on ${{ matrix.os }}
         with:
           node-version: ${{ matrix.node-version }}
       - run: sh sql/init-mysql.sh
@@ -130,7 +122,6 @@ jobs:
           choco install mysql
       - uses: actions/checkout@v6
       - uses: actions/setup-node@v6
-        name: Node ${{ matrix.node-version }}
         with:
           node-version: ${{ matrix.node-version }}
       - run: sh sql/init-mysql.sh
