+++ .github/workflows/publish.yml
@@ -6,7 +6,6 @@ on:
 
 env:
   CI: true
-  node-version: 22
 
 jobs:
   build:
@@ -17,8 +16,6 @@ jobs:
       - run: sudo /etc/init.d/mysql start
       - uses: actions/checkout@v6
       - uses: actions/setup-node@v6
-        with:
-          node-version: ${{ env.node-version }}
       - run: sh sql/init-mysql.sh
       - run: npm install
       - run: npm test
@@ -31,16 +28,12 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/setup-node@v6
-        name: Node ${{ env.node-version }}
         with:
-          node-version: ${{ env.node-version }}
           registry-url: https://registry.npmjs.org/
-
       - uses: actions/checkout@v6
         with:
           fetch-depth: 0
           # fetch-depth 0 needed by GitHub Release
-
       - name: publish to NPM
         run: |
           VERSION=$(node -e 'console.log(require("./package.json").version)')
@@ -61,7 +54,6 @@ jobs:
       - uses: actions/checkout@v6
       - uses: actions/setup-node@v6
         with:
-          node-version: ${{ env.node-version }}
           registry-url: https://npm.pkg.github.com/
           scope: '@nictool'
       - run: npm publish
